+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-30 10:26:06
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-07-30 12:01:45
+// Commit: perf(state): apply RepaintBoundary on animated state widgets
 
 class stateComponentState {
-  final String commitRef = '3228';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-07-30T10:26:06+05:00');
+  final String commitRef = '3236';
+  final String lastMessage = 'perf(state): apply RepaintBoundary on animated state widgets';
+  final DateTime timestamp = DateTime.parse('2026-07-30T12:01:45+05:00');
 }